Split global.go
